[refactor] #24 - 파일이름에 view 삭제
