+++ src/frontend/src/components/features/ChallengeDetail/ChallengeDetail.jsx
