+++ src/frontend/src/components/features/ChallengeDetail/ChallengeDetail.module.css
